Moved the extensions to the approppriate class.
